% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjf_baixar_cjsg.R
\name{cjf_baixar_cjsg}
\alias{cjf_baixar_cjsg}
\title{Baixar jurisprudência da justiça federal}
\usage{
cjf_baixar_cjsg(
  livre = "",
  aspas = FALSE,
  numero = "",
  classe = "",
  relator = "",
  revisor = "",
  relator_convocado = "",
  relator_acordao = "",
  ementa_decisao = "",
  referencia_legislativa = "",
  data_inicial = "",
  data_final = "",
  tipo = c("julgamento", "publicacao"),
  tribunal = NULL,
  diretorio = "."
)
}
\arguments{
\item{livre}{Busca livre}

\item{aspas}{Colocar entre aspas?}

\item{numero}{Número}

\item{classe}{Classe}

\item{relator}{Relator}

\item{revisor}{Revisor}

\item{relator_convocado}{Relator convocado}

\item{relator_acordao}{Relacor do acórdão}

\item{ementa_decisao}{Busca na ementa ou decisão}

\item{referencia_legislativa}{Referência legislativa}

\item{data_inicial}{Data inicial no formato "dd/mm/aaaa"}

\item{data_final}{Data final no formato "dd/mm/aaaa"}

\item{tipo}{Datas por "julgamento" ou "publicacao". Default: julgamento}

\item{tribunal}{informar tribunal, e.g, "trf1" ou "stf"}

\item{diretorio}{Diretório onde armazenar arquivos}
}
\value{
html
}
\description{
Baixar jurisprudência da justiça federal
}
