% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjf_ler_cjsg.R
\name{cjf_ler_cjsg}
\alias{cjf_ler_cjsg}
\title{Lê jurisprudência baixada por cjf_baixar_cjsg}
\usage{
cjf_ler_cjsg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar diretório se não forneceu arquivos}
}
\value{
Tibble
}
\description{
Lê jurisprudência baixada por cjf_baixar_cjsg
}
